refactor: try to make pass Sonar test check
