+++ README.md
@@ -10,5 +10,5 @@ cd C:/Путь/До/Папки/С этой/Программой
 Также можно запустить файл с тестами:
 ./test.sh
 В связи с тем, что Sonar за отсутствие тестов бьёт палками, добавляю сюда команды для создания тестов:
-kotlinc app.kt tests.kt -classpath "libs/junit-platform-console-standalone.jar;libs/kotlinx-cli-jvm-0.3.6.jar" -d test.jar
+kotlinc app.kt -classpath "libs/junit-platform-console-standalone.jar;libs/kotlinx-cli-jvm-0.3.6.jar" -d test.jar
 java -jar libs/junit-platform-console-standalone.jar --class-path test.jar;libs/kotlin-stdlib.jar --scan-class-path
